# Avoid intermediate JSON object allocations when generating JSON strings