--- a/protobuf/lib/meta.dart
+++ b/protobuf/lib/meta.dart
@@ -60,6 +60,7 @@ const GeneratedMessage_reservedNames = <String>[
   'writeToCodedBufferWriter',
   'writeToJson',
   'writeToJsonMap',
+  'writeToJsonSink',
   r'$_ensure',
   r'$_get',
   r'$_getI64',