--- a/protobuf/lib/src/protobuf/builder_info.dart
+++ b/protobuf/lib/src/protobuf/builder_info.dart
@@ -41,17 +41,21 @@ class BuilderInfo {
 
   List<FieldInfo>? _sortedByTag;
 
-  // For well-known types.
-  final Object? Function(GeneratedMessage message, TypeRegistry typeRegistry)?
-      toProto3Json;
+  /// JSON generator for well-known types.
+  final void Function(
+          GeneratedMessage msg, TypeRegistry typeRegistry, JsonSink jsonSink)?
+      writeToProto3JsonSink;
+
+  /// JSON parser for well-known types.
   final Function(GeneratedMessage targetMessage, Object json,
       TypeRegistry typeRegistry, JsonParsingContext context)? fromProto3Json;
+
   final CreateBuilderFunc? createEmptyInstance;
 
   BuilderInfo(String? messageName,
       {PackageName package = const PackageName(''),
       this.createEmptyInstance,
-      this.toProto3Json,
+      this.writeToProto3JsonSink,
       this.fromProto3Json})
       : qualifiedMessageName = '${package.prefix}$messageName';