Use jsontool library to avoid generating intermediate JSON objects when
generating JSON strings from messages.

Original proto3 code moved from #670. Generator for other JSON format is
updated to use jsontool as well.

Splits JSON parsers and generators into separate files.

Benchmark results below are noisy, I will try to find a way to reduce noise.

## AOT runtime

|                    | Before    | After    | Diff
|--------------------|-----------|----------|-
|ToJsonObject        | 10304 μs  | 11175 μs | +871 μs, +8.45%
|ToJsonString        | 35030 μs  | 25946 μs | -9084 μs, -25.93%
|ToProto3JsonObject  | 10812 μs  | 10824 μs | +12 μs, +0.11%
|ToProto3JsonString  | 35906 μs  | 26758 μs | -9148 μs, -25.48%

## JS runtime

|                    | Before    | After    | Diff
|--------------------|-----------|----------|-
|ToJsonObject        | 19637 μs  | 23313 μs | +3676 μs, +18.72%
|ToJsonString        | 71535 μs  | 45477 μs | -26058 μs, -36.43%
|ToProto3JsonObject  | 22318 μs  | 20323 μs | -1995 μs, -8.94%
|ToProto3JsonString  | 71642 μs  | 50000 μs | -21642 μs, -30.21%

## Benchmark binary size

|    | Before          | After          | Diff
|----|-----------------|----------------|-
|AOT | 6,990,992 bytes | 7,010,576 bytes| +19,584 bytes, +0.28%
|JS  | 338,060 bytes   | 339,829 bytes  | +1,769 bytes +0.52%