@lrhn 2022-09-19
The jsontool package is a third-party package which is not supported by the Dart team.
Would you consider forking or vendoring it?