@osa1 2023-10-27
We should add `GeneratedMessage` members to generate encoded JSON, ready for writing to a socket.

Secondly, we can keep the current `toJsonObject` implementation as it's faster on both AOT and JS.